client/requirements.txt
pyaudiowpatch>=0.2.12
websocket-client>=1.7.0
numpy>=1.26.0
